First steps (tested)
Here is the initial .c program that I made (mostly from LabJack example code) for reading in voltages and saving them to a .csv file. I just tested it in Nicolette's room with a labjack and it was successful. Next steps should be adding a loop (re-checking values 100 per second) and considering 16-bit data.